requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
flatbuffers==25.12.19
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.68.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
matplotlib==3.10.8
mediapipe==0.10.9
narwhals==2.18.0
numpy==1.26.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
pandas==2.3.3
pillow==12.1.1
protobuf==3.20.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sounddevice==0.5.5
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.5.5
typing-inspection==0.4.2
tzdata==2025.3
websockets==16.0
